face door opening system 111  some error/requirements.txt
opencv-python>=4.5.0
face-recognition>=1.3.0
numpy>=1.19.0
pyttsx3>=2.90
RPi.GPIO>=0.7.0
flask>=2.0.0
Pillow>=8.0.0